Python: Adds pytest option to exclude the Redis modules tests from the default test run
