+++ python/pytest.ini
@@ -1,3 +1,4 @@
 [pytest]
 markers =
     smoke_test: mark a test as a build verification testing.
+addopts = -k "not test_redis_modules.py"
